Move NGS functions to single file; add analysis type from filename test.
