--- test_data/input_data/CLINICAL/NGS/pmc_test.seg.sha1
@@ -1,1 +1,0 @@
-b4f261f555757cb1eeaa06c14efdc76b1d7c25d5  pmc_test.seg
