+++ tests/sources2csr/test_derived_values.py
@@ -3,13 +3,10 @@
 
 """Tests for the derived diagnosis aggregates.
 """
-from collections import Counter
-
 import pytest
 from csr.csr import CentralSubjectRegistry, Individual, Diagnosis, Biosource, Biomaterial
 
-from sources2csr.helper_variables import add_derived_values, add_ngs_data
-from sources2csr.ngs import NGS, LibraryStrategy, AnalysisType
+from sources2csr.derived_values import add_derived_values
 
 
 @pytest.fixture
@@ -61,42 +58,3 @@ def test_diagnosis_aggregates_with_missing_data(registry_with_missing_diagnosis_
     for individual in subject_registry.individuals:
         assert individual.age_first_diagnosis == expected_age[individual.individual_id]
         assert individual.diagnosis_count == expected_counts[individual.individual_id]
-
-
-@pytest.fixture
-def registry_with_biosources_and_biomaterials() -> CentralSubjectRegistry:
-    individuals = [
-        Individual(individual_id='P1', birth_date='1950-05-05')
-    ]
-    diagnoses = [
-        Diagnosis(diagnosis_id='D1', individual_id='P1', diagnosis_date='2010-03-31')
-    ]
-    biosources = [
-        Biosource(biosource_id='BS1', individual_id='P1', diagnosis_id='D1')
-    ]
-    biomaterials = [
-        Biomaterial(biomaterial_id='BM1', src_biosource_id='BS1'),
-        Biomaterial(biomaterial_id='BM2', src_biosource_id='BS1')
-    ]
-
-    return CentralSubjectRegistry(individuals=individuals, diagnoses=diagnoses, biosources=biosources,
-                                  biomaterials=biomaterials)
-
-
-def test_biomaterial_library_strategy_and_analysis_type(registry_with_biosources_and_biomaterials):
-    ngs_data = [
-        NGS(biosource_id='BS1', biomaterial_id='BM1', library_strategy=LibraryStrategy.CNV),
-        NGS(biosource_id='BS1', biomaterial_id='BM1', library_strategy=LibraryStrategy.CNV),
-        NGS(biosource_id='BS1', biomaterial_id='BM1', library_strategy=LibraryStrategy.SNV),
-        NGS(biosource_id='BS1', biomaterial_id='BM2', library_strategy=LibraryStrategy.CNV,
-            analysis_type=AnalysisType.WGS),
-        NGS(biosource_id='BS2', biomaterial_id='BM2', library_strategy=LibraryStrategy.CNV,
-            analysis_type=AnalysisType.WXS),
-    ]
-    subject_registry = add_ngs_data(registry_with_biosources_and_biomaterials, set(ngs_data))
-
-    assert len(subject_registry.biomaterials) == 2
-    assert Counter(subject_registry.biomaterials[0].library_strategy) == Counter(['CNV', 'SNV'])
-    assert subject_registry.biomaterials[1].library_strategy == ['CNV']
-    assert subject_registry.biomaterials[0].analysis_type == []
-    assert subject_registry.biomaterials[1].analysis_type == ['WGS']
